Refactor KFDataStore to become a class cluster
The underlying core data stack will depend on which internal class is
used.

This allows a greater amount of flexibility and higher performance
depending on the applications needs.

The existing core data stack using many child contexts is now
abandoned, this was a terrible way of dealing with core data. Any
read/writes in a background can cause the parent context to block then
when the a managed object context on the main queue fetches or saves
data it can result in a lock.